Corrige el tipo de las tasas y endurece las pruebas de exportación
Las tasas de resolución y de sin contacto salían como float de round(),
así que el snapshot del informe guardaba 100.0 en vez de 100. Se castean
a entero: son porcentajes y se congelan tal cual en el JSON publicado.

Las pruebas de CSV comparaban contra el texto crudo, pero fputcsv
entrecomilla cualquier campo con espacios. Ahora se interpreta el CSV de
vuelta con str_getcsv en lugar de buscar subcadenas.

+++ app/Support/MetricasTecnico.php
@@ -103,8 +103,10 @@ class MetricasTecnico
             'enviados_campo' => (clone $this->base())->whereHas('ordenesTrabajo')->count(),
             'inmediatos' => (clone $this->base())->where('criticidad', Soporte::CRITICIDAD_INMEDIATA)->count(),
 
-            'tasa_resolucion' => $total > 0 ? round($resueltos / $total * 100) : null,
-            'tasa_sin_contacto' => $total > 0 ? round($sinContacto / $total * 100) : null,
+            // Enteros a propósito: son porcentajes, y además se guardan tal cual
+            // en el snapshot JSON del informe publicado.
+            'tasa_resolucion' => $total > 0 ? (int) round($resueltos / $total * 100) : null,
+            'tasa_sin_contacto' => $total > 0 ? (int) round($sinContacto / $total * 100) : null,
             'promedio_resolucion' => $promedio ? (int) round($promedio) : null,
 
             'dias_trabajados' => count($porDia),
